Added support of text styles in messages (#104)
* Added support of text styles in messages

* Fixed SkipSelfSent and OnlySelfSent methods

+++ Sources/Falko.Talkie.Platforms.Telegram/Connections/TelegramSignalConnection.cs
@@ -46,7 +46,7 @@ public sealed class TelegramSignalConnection(ISignalFlow flow,
 
     protected override async Task WhenExecutingAsync(CancellationToken cancellationToken)
     {
-        long offset = 0;
+        int offset = 0;
 
         while (cancellationToken.IsCancellationRequested is false)
         {
